% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_funcoes_graficos.R
\name{grafico_medias_fatorial}
\alias{grafico_medias_fatorial}
\title{Grafico de medias com letras de comparacao}
\usage{
grafico_medias_fatorial(
  dados,
  resposta,
  fator_interesse,
  bloco = NULL,
  fatores,
  dic_vars = NULL,
  label_type = c("label", "sigla", "description", "var"),
  digitos = 2
)
}
\arguments{
\item{dados}{\code{data.frame} com os dados experimentais.}

\item{resposta}{Nome da variavel resposta.}

\item{fator_interesse}{Nome do fator para comparacao de medias.}

\item{bloco}{Nome do bloco (DBC). Use \code{NULL} para DIC.}

\item{fatores}{Vetor de nomes dos fatores do modelo.}

\item{dic_vars}{Dicionario opcional com colunas \code{var}, \code{sigla}, \code{label}, \code{description}.}

\item{label_type}{Tipo de rotulo da variavel resposta.}

\item{digitos}{Numero de casas decimais.}
}
\value{
Objeto \code{ggplot2}.
}
\description{
Grafico de medias com letras de comparacao
}
