% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranova-package.R
\docType{package}
\name{ranova-package}
\alias{ranova}
\alias{ranova-package}
\title{ranova: Analise de Variancia para Experimentos}
\description{
Ferramentas para ajustar experimentos simples e em esquema fatorial (DIC e
DBC), em parcelas subdivididas e subsubdivididas, gerar ANOVA, testes de
medias, desdobramentos, diagnostico de pressupostos, tabelas e graficos
para relatorios. O motor principal e \code{\link[=ranova_ajuste]{ranova_ajuste()}}, com
\code{\link[=ranova_anova]{ranova_anova()}} e \code{\link[=ranova_medias]{ranova_medias()}}.
}
\author{
\strong{Maintainer}: Marlenildo Ferreira Melo \email{marlenildo@gmail.com}

}
\keyword{internal}
