% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_testes_medias.R
\docType{data}
\name{TESTES_MEDIAS}
\alias{TESTES_MEDIAS}
\title{Testes de medias disponiveis}
\format{
Vetor nomeado de caracteres.
}
\usage{
TESTES_MEDIAS
}
\description{
Nomes aceitos pelo argumento \code{teste} de \code{\link[=letras_teste]{letras_teste()}} e
\code{\link[=ranova_medias]{ranova_medias()}}, com o rotulo em portugues.
}
\keyword{datasets}
